+++ lib/puree/collection.rb
@@ -1,25 +1,31 @@
 module Puree
 
-  # Collection resource
+  # Collection of resources
   #
-  class Collection < Resource
+  class Collection
 
-    # @param api [Symbol]
+    # @param resource [Symbol]
     # @param endpoint [String]
     # @param optional username [String]
     # @param optional password [String]
-    def initialize(api: nil,
+    # @param optional basic_auth [Boolean]
+    def initialize(resource: nil,
                    endpoint: nil,
                    username: nil,
-                   password: nil)
-      super(api: api,
-            endpoint: endpoint,
-            username: username,
-            password: password)
+                   password: nil,
+                   basic_auth: nil)
+      @resource_type = resource
+      @api_map = Puree::Map.new.get
+      @endpoint = endpoint.nil? ? Puree.endpoint : endpoint
+      @basic_auth = basic_auth.nil? ? Puree.basic_auth : basic_auth
+      if @basic_auth === true
+        @username = username.nil? ? Puree.username : username
+        @password = password.nil? ? Puree.password : password
+      end
       @uuids = []
     end
 
-    # Get
+    # Gets an array of objects of resource type specified in constructor
     #
     # @param optional limit [Integer]
     # @param optional offset [Integer]
@@ -27,27 +33,20 @@ module Puree
     # @param optional created_end [String]
     # @param optional modified_start [String]
     # @param optional modified_end [String]
-    # @return [HTTParty::Response]
+    # @param optional full [Boolean]
+    # @return [Array<Object>]
     def get(
             limit:            20,
             offset:           0,
             created_start:    nil,
             created_end:      nil,
             modified_start:   nil,
-            modified_end:     nil)
-
-      missing = missing_credentials
-      if !missing.empty?
-        missing.each do |m|
-          puts "#{self.class.name}" + '#' + "#{__method__} missing #{m}"
-        end
-        exit
-      end
-      # strip any trailing slash
-      @endpoint = @endpoint.sub(/(\/)+$/, '')
-      @auth = Base64::strict_encode64(@username + ':' + @password)
+            modified_end:     nil,
+            full:             true
+    )
 
       @options = {
+          basic_auth:       @basic_auth,
           latest_api:       true,
           resource_type:    @resource_type.to_sym,
           rendering:        :system,
@@ -56,12 +55,29 @@ module Puree
           created_start:    created_start,
           created_end:      created_end,
           modified_start:   modified_start,
-          modified_end:     modified_end
+          modified_end:     modified_end,
+          full:             full
       }
-      headers = {
-          'Accept' => 'application/xml',
-          'Authorization' => 'Basic ' + @auth
-      }
+
+      missing = missing_credentials
+      if !missing.empty?
+        missing.each do |m|
+          puts "#{self.class.name}" + '#' + "#{__method__} missing #{m}"
+        end
+        exit
+      end
+
+      # strip any trailing slash
+      @endpoint = @endpoint.sub(/(\/)+$/, '')
+
+      headers = {}
+      headers['Accept'] = 'application/xml'
+
+      if @options[:basic_auth] === true
+        @auth = Base64::strict_encode64(@username + ':' + @password)
+        headers['Authorization'] = 'Basic ' + @auth
+      end
+
       query = {}
 
       query['rendering'] = @options[:rendering]
@@ -91,7 +107,23 @@ module Puree
       end
 
       @response = HTTParty.get(build_url, query: query, headers: headers)
+
+      if @options[:full]
+        collect_resource
+      else
+        data = []
+        uuid.each do |u|
+          o = {}
+          o['uuid'] = u
+          data << o
+        end
+        data
+      end
+
     end
+
+
+    private
 
 
     # Array of UUIDs
@@ -103,9 +135,31 @@ module Puree
     end
 
 
+    def collect_resource
+      data = []
+      resource_class = 'Puree::' + @resource_type.to_s.capitalize
 
-    private
-
+      if @options[:basic_auth] === true
+        r = Object.const_get(resource_class).new endpoint:   @endpoint,
+                                                 username:   @username,
+                                                 password:   @password
+      else
+        r = Object.const_get(resource_class).new endpoint:   @endpoint,
+                                                 basic_auth: false
+      end
+      # whitelist symbol
+      if @api_map[:resource_type].has_key?(@resource_type)
+        uuid.each do |u|
+          record = r.find uuid: u
+          # puts record
+          data << record
+        end
+        data
+      else
+        puts 'Invalid resource class'
+        exit
+      end
+    end
 
     def collect_uuid
       path = '//renderedItem/@renderedContentUUID'
@@ -113,8 +167,52 @@ module Puree
       xpath_result.each { |i| @uuids << i.text.strip }
     end
 
+    def service_name
+      resource_type = @options[:resource_type]
+      @api_map[:resource_type][resource_type][:service]
+    end
+
+    def service_response_name
+      resource_type = @options[:resource_type]
+      @api_map[:resource_type][resource_type][:response]
+    end
+
+    def build_url
+      service = service_name
+      if @options[:latest_api] === false
+        service_api_mode = service
+      else
+        service_api_mode = service + '.current'
+      end
+      @endpoint + '/' + service_api_mode
+    end
+
+    def xpath_query(path)
+      xml = @response.body
+      doc = Nokogiri::XML xml
+      doc.remove_namespaces!
+      doc.xpath path
+    end
+
+    def missing_credentials
+      missing = []
+      if @endpoint.nil?
+        missing << 'endpoint'
+      end
+
+      if @options[:basic_auth] === true
+        if @username.nil?
+          missing << 'username'
+        end
+        if @password.nil?
+          missing << 'password'
+        end
+      end
+
+      missing
+    end
+
     alias :find :get
 
   end
-
 end