+++ lib/puree/configuration.rb
@@ -2,15 +2,11 @@ module Puree
 
   module Configuration
 
-    class << self
-
-      attr_accessor :endpoint, :username, :password
+      attr_accessor :endpoint, :username, :password, :basic_auth
 
       def configure
         yield self
       end
-
-    end
 
   end