+++ lib/puree/journal.rb
@@ -7,12 +7,14 @@ module Puree
     # @param endpoint [String]
     # @param optional username [String]
     # @param optional password [String]
-    def initialize(endpoint: nil, username: nil, password: nil)
+    def initialize(endpoint: nil, username: nil, password: nil, basic_auth: nil)
       super(api: :journal,
             endpoint: endpoint,
             username: username,
-            password: password)
+            password: password,
+            basic_auth: basic_auth)
     end
+
 
     # All metadata
     #