+++ lib/puree/map.rb
@@ -21,13 +21,19 @@ module Puree
             family: 'dk.atira.pure.modules.datasets.external.model.dataset.DataSet'
           },
           download: {
-              service: 'downloadcount',
-              response: 'GetDownloadCountResponse'
+            service: 'downloadcount',
+            response: 'GetDownloadCountResponse'
           },
+          server: {
+            service: 'servermeta',
+            response: 'GetServerMetaResponse'
+          },
+
         }
       }
 
       add_convention
+      # add_family
     end
 
     # Pure API map
@@ -36,6 +42,7 @@ module Puree
     def get
       @api_map
     end
+
 
 
     private
@@ -48,6 +55,10 @@ module Puree
         @api_map[:resource_type][c.to_sym] = resource_type
       end
     end
+
+    # def add_family
+    #   @api_map[:resource_type][:event][:family] = 'dk.atira.pure.api.shared.model.event.Event'
+    # end
 
 
   end