+++ lib/puree/organisation.rb
@@ -7,12 +7,15 @@ module Puree
     # @param endpoint [String]
     # @param optional username [String]
     # @param optional password [String]
-    def initialize(endpoint: nil, username: nil, password: nil)
+    def initialize(endpoint: nil, username: nil, password: nil, basic_auth: nil)
       super(api: :organisation,
             endpoint: endpoint,
             username: username,
-            password: password)
+            password: password,
+            basic_auth: basic_auth)
     end
+
+
 
     # Address
     #
@@ -112,8 +115,6 @@ module Puree
       o
     end
 
-
-    private
 
 
     # Organisation