+++ spec/collection.rb
@@ -3,7 +3,13 @@ require 'spec_helper'
 describe 'Collection' do
 
   it '#new' do
-    p = Puree::Collection.new(api: :dataset)
+    endpoint = ENV['PURE_ENDPOINT']
+    username = ENV['PURE_USERNAME']
+    password = ENV['PURE_PASSWORD']
+    p = Puree::Collection.new(resource: :dataset,
+                              endpoint: endpoint,
+                              username: username,
+                              password: password)
     expect(p).to be_an_instance_of Puree::Collection
   end
 
@@ -12,15 +18,15 @@ describe 'Collection' do
       endpoint = ENV['PURE_ENDPOINT']
       username = ENV['PURE_USERNAME']
       password = ENV['PURE_PASSWORD']
-      @p = Puree::Collection.new(api: :dataset,
+      @p = Puree::Collection.new(resource: :dataset,
                                  endpoint: endpoint,
                                  username: username,
                                  password: password)
-      @p.find limit: 5
+      @metadata = @p.find limit: 5
     end
 
-    it '#UUID' do
-      expect(@p.uuid).to be_an_instance_of(Array)
+    it '@metadata' do
+      expect(@metadata).to be_an_instance_of(Array)
     end
 
   end